engine: extract esm/{toast_tile,triangle}.js
